+++ src/lib/components/Sidebar/index.ts
@@ -1,4 +1,4 @@
 export { default as Sidebar } from './Sidebar.svelte'
 export { default as SidebarTrigger } from './SidebarTrigger.svelte'
 export type { SidebarTriggerProps } from './sidebar-trigger.types.js'
-export type { SidebarProps, SidebarMode } from './sidebar.types.js'
+export type { SidebarProps, SidebarMode, SidebarApi } from './sidebar.types.js'